[refactor]: Move data logic out of WidgetView
